Move test for font color import to new file
