+++ xlsx/tests/test.rs
@@ -479,47 +479,3 @@ fn test_workbook_theme_colors() {
     let style_b15 = model.get_style_for_cell(0, 15, 2).unwrap();
     assert_eq!(style_b15.fill.fg_color.as_deref(), Some("#C9211E"));
 }
-
-#[test]
-// A <font> with no <color> child is semantically equivalent to <color auto="1"/> in OOXML —
-// both mean "use the automatic/default color." The importer must surface this as
-// Font.color = None so callers can distinguish it from an explicit <color rgb="FF000000"/>.
-//
-// PRODUCT_SUM.xlsx contains three fonts: two with <color theme="1"/> (sz=11) and one with
-// no <color> child (sz=8, the only font with that size).
-fn test_font_with_no_color_element_is_none() {
-    let model = load_from_xlsx("tests/calc_tests/PRODUCT_SUM.xlsx", "en", "UTC", "en").unwrap();
-    let fonts = &model.workbook.styles.fonts;
-
-    let no_color_font = fonts
-        .iter()
-        .find(|f| f.sz == 8)
-        .expect("expected a font with sz=8 (the one with no <color> element)");
-    assert_eq!(
-        no_color_font.color, None,
-        "font with no <color> element should import as None, not Some(\"#000000\")"
-    );
-
-    let themed_fonts: Vec<_> = fonts.iter().filter(|f| f.sz == 11).collect();
-    assert!(!themed_fonts.is_empty(), "expected at least one sz=11 font");
-    for font in themed_fonts {
-        assert!(
-            font.color.is_some(),
-            "font with <color theme=...> should resolve to Some(rgb), got None"
-        );
-    }
-
-    // Round-trip: the no-<color> font must not gain an explicit color on export.
-    let temp_file_name = "temp_file_test_font_no_color.xlsx";
-    save_to_xlsx(&model, temp_file_name).unwrap();
-    let reloaded = load_from_xlsx(temp_file_name, "en", "UTC", "en").unwrap();
-    fs::remove_file(temp_file_name).unwrap();
-    let reloaded_no_color = reloaded
-        .workbook
-        .styles
-        .fonts
-        .iter()
-        .find(|f| f.sz == 8)
-        .expect("expected sz=8 font after round-trip");
-    assert_eq!(reloaded_no_color.color, None);
-}
